# docs/swagger.yaml
basePath: /v1
definitions:
...
  user.Resource:
    properties:
      ID:
        type: integer
      email:
        type: string
      password:
        type: string
    type: object
...
host: cinema-alb-dev-o81jt53c-906642332.us-east-1.elb.amazonaws.com:8085
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample cinetickets server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Cinetickets API
  version: "1.0"
paths:
...
  /signin:
    post:
      consumes:
      - application/json
      description: Signin
      parameters:
      - description: Email and password
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/user.Resource'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: "Bad Request"
        "422":
          description: "Unprocessable Entity"
        "500":
          description: "Internal Server Error"
      summary: Signin
      tags:
      - Users
  /signup:
    post:
      consumes:
      - application/json
      description: Signup
      parameters:
      - description: Email and password
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/user.Resource'
      produces:
      - application/json
      responses:
        "201":
          description: ""
        "400":
          description: "Bad Request"
        "422":
          description: "Unprocessable Entity"
        "500":
          description: "Internal Server Error"
      summary: Signup
      tags:
      - Users
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"